pyproject.toml
[project]
name = "03-expense-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
